some changes

+++ include/order.h
@@ -7,25 +7,25 @@ class Order
 {
 public:
     Order() = default;
-    Order(Product **order_list, int price = 0,
+    Order(Product **order_list, float price = 0,
           unsigned int o_list_physical_size = 0, unsigned int o_list_logical_size = 0);
     Order(const Order &other);
     Order(Order &&other); //
     ~Order();
 
-    bool setPrice(int price_of_order);
+    bool setPrice(float price_of_order);
     bool setProductList(Product **order_list);
     bool setPhysicalSize(unsigned int physical_size);
     bool setLogicalSize(unsigned int logical_size);
     bool addProductToOrderList(Product *new_product);
-    int getPrice() const;
+    float getPrice() const;
     Product **getList() const;
     unsigned int getPhysicalSize() const;
     unsigned int getLogicalSize() const;
 
 private:
     void resizeOrderList();
-    int price_of_order = 0;
+    float price_of_order = 0;
     Product **order_products = nullptr; //pointer to array of products we have already bought.
     unsigned int o_list_physical_size = 0;
     unsigned int o_list_logical_size = 0;
